# ai-version/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "To Do List API",
    "version": "1.0.0",
    "description": "A simple in-memory To Do List API built with Express.js."
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "tags": [
    {
      "name": "System",
      "description": "API status and metadata endpoints."
    },
    {
      "name": "Tasks",
      "description": "CRUD, search, filter, and statistics endpoints for tasks."
    }
  ],
  "paths": {
    "/": {
      "get": {
        "tags": [
          "System"
        ],
        "summary": "Get API information",
        "responses": {
          "200": {
            "description": "API information returned successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiInfoResponse"
                }
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "tags": [
          "System"
        ],
        "summary": "Check API health",
        "responses": {
          "200": {
            "description": "API is running.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/tasks": {
      "get": {
        "tags": [
          "Tasks"
        ],
        "summary": "List tasks",
        "description": "Returns all tasks. Supports optional filtering by done status and searching by title.",
        "parameters": [
          {
            "name": "done",
            "in": "query",
            "required": false,
            "description": "Filter tasks by done status.",
            "schema": {
              "type": "boolean"
            },
            "examples": {
              "done": {
                "value": true
              },
              "notDone": {
                "value": false
              }
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "description": "Case-insensitive search against task titles.",
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "example": "swagger"
          }
        ],
        "responses": {
          "200": {
            "description": "Tasks returned successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskListResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      },
      "post": {
        "tags": [
          "Tasks"
        ],
        "summary": "Create a task",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTaskRequest"
              },
              "examples": {
                "default": {
                  "value": {
                    "title": "Buy groceries",
                    "done": false
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Task created successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
    "/tasks/stats": {
      "get": {
        "tags": [
          "Tasks"
        ],
        "summary": "Get task statistics",
        "responses": {
          "200": {
            "description": "Task statistics returned successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/tasks/{id}": {
      "get": {
        "tags": [
          "Tasks"
        ],
        "summary": "Get a task by ID",
        "parameters": [
          {
            "$ref": "#/components/parameters/TaskId"
          }
        ],
        "responses": {
          "200": {
            "description": "Task returned successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "put": {
        "tags": [
          "Tasks"
        ],
        "summary": "Replace a task",
        "parameters": [
          {
            "$ref": "#/components/parameters/TaskId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTaskRequest"
              },
              "examples": {
                "default": {
                  "value": {
                    "title": "Buy groceries and cook dinner",
                    "done": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Task updated successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "patch": {
        "tags": [
          "Tasks"
        ],
        "summary": "Partially update a task",
        "parameters": [
          {
            "$ref": "#/components/parameters/TaskId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchTaskRequest"
              },
              "examples": {
                "titleOnly": {
                  "value": {
                    "title": "Buy groceries today"
                  }
                },
                "doneOnly": {
                  "value": {
                    "done": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Task updated successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "delete": {
        "tags": [
          "Tasks"
        ],
        "summary": "Delete a task",
        "parameters": [
          {
            "$ref": "#/components/parameters/TaskId"
          }
        ],
        "responses": {
          "200": {
            "description": "Task deleted successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    }
  },
  "components": {
    "parameters": {
      "TaskId": {
        "name": "id",
        "in": "path",
        "required": true,
        "description": "Positive integer task ID.",
        "schema": {
          "type": "integer",
          "minimum": 1
        },
        "example": 1
      }
    },
    "responses": {
      "BadRequest": {
        "description": "The request is invalid.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "NotFound": {
        "description": "The requested resource was not found.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    },
    "schemas": {
      "Task": {
        "type": "object",
        "required": [
          "id",
          "title",
          "done"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "title": {
            "type": "string",
            "example": "Buy groceries"
          },
          "done": {
            "type": "boolean",
            "example": false
          }
        }
      },
      "CreateTaskRequest": {
        "type": "object",
        "required": [
          "title"
        ],
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "example": "Buy groceries"
          },
          "done": {
            "type": "boolean",
            "default": false,
            "example": false
          }
        }
      },
      "UpdateTaskRequest": {
        "type": "object",
        "required": [
          "title",
          "done"
        ],
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "example": "Buy groceries and cook dinner"
          },
          "done": {
            "type": "boolean",
            "example": true
          }
        }
      },
      "PatchTaskRequest": {
        "type": "object",
        "minProperties": 1,
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "example": "Buy groceries today"
          },
          "done": {
            "type": "boolean",
            "example": true
          }
        }
      },
      "SuccessResponse": {
        "type": "object",
        "required": [
          "success",
          "message"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "Operation completed successfully."
          }
        }
      },
      "TaskResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SuccessResponse"
          },
          {
            "type": "object",
            "required": [
              "data"
            ],
            "properties": {
              "data": {
                "$ref": "#/components/schemas/Task"
              }
            }
          }
        ]
      },
      "TaskListResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SuccessResponse"
          },
          {
            "type": "object",
            "required": [
              "data"
            ],
            "properties": {
              "data": {
                "type": "object",
                "required": [
                  "count",
                  "tasks"
                ],
                "properties": {
                  "count": {
                    "type": "integer",
                    "example": 3
                  },
                  "tasks": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Task"
                    }
                  }
                }
              }
            }
          }
        ]
      },
      "StatsResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SuccessResponse"
          },
          {
            "type": "object",
            "required": [
              "data"
            ],
            "properties": {
              "data": {
                "type": "object",
                "required": [
                  "total",
                  "done",
                  "notDone"
                ],
                "properties": {
                  "total": {
                    "type": "integer",
                    "example": 3
                  },
                  "done": {
                    "type": "integer",
                    "example": 1
                  },
                  "notDone": {
                    "type": "integer",
                    "example": 2
                  }
                }
              }
            }
          }
        ]
      },
      "ApiInfoResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SuccessResponse"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "type": "object"
              }
            }
          }
        ]
      },
      "HealthResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SuccessResponse"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "type": "object",
                "properties": {
                  "status": {
                    "type": "string",
                    "example": "ok"
                  }
                }
              }
            }
          }
        ]
      },
      "ErrorResponse": {
        "type": "object",
        "required": [
          "success",
          "message"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "Task with ID 99 was not found."
          },
          "details": {
            "type": "object",
            "nullable": true
          }
        }
      }
    }
  }
}